Remove eth transfer events table and simplify eth indexing flow
